scm_backend/requirements.txt
django ~= 4.1.1
djangorestframework ~= 3.13.1
django-cors-headers ~= 3.13.0
rake-nltk ~= 1.0.6
nltk ~= 3.7
pandas ~= 1.5.0
openpyxl ~= 3.0.10
spacy ~= 3.4.1
pytextrank ~= 3.2.4
keybert ~= 0.7.0
yake ~= 0.4.8
monkeylearn ~= 3.6.0
textrazor ~= 1.4.0
scikit-learn ~= 1.2.0
Unidecode ~= 1.3.6
matplotlib ~= 3.6.1
gensim ~= 4.2.0
nltk ~= 3.7
python-dotenv ~= 0.21.0
scrapy ~= 2.7.1
cloudscraper ~= 1.2.66
configparser ~= 5.3.0